refactored uist http into v1, removed client module and tried to tidy
