+++ rotala/src/http/mod.rs
@@ -1,2 +1,2 @@
-pub mod jura;
-pub mod uist;
+pub mod jura_v1;
+pub mod uist_v1;
